Add files via upload
This is a EDA process of a dataset that contains demographics about certain customers and based on whether that customer clicked the ad or not a purchased was made . 

Also a classification algorithm was used to predict if the demographic column Age could predict the purchased or not. 

This data set contains the following features: 'User ID': unique identification for consumer, 'Age': customer age in years, 'Estimated Salary': Avg. Income of consumer, 'Gender': Whether consumer was male or female and 'Purchased': 0 or 1 indicated clicking on Ad.